Extract useIsOnline hook
